+++ src/test/php/peer/http/HttpUriTest.php
@@ -258,12 +258,26 @@ class HttpUriTest extends \PHPUnit_Framework_TestCase
      * @since  2.0.0
      * @test
      */
-    public function transposingToHttpsLeavesEverythingExceptScheme()
+    public function transposingToHttpsLeavesEverythingExceptSchemeAndPort()
     {
         $this->assertEquals(
-                'https://example.net:8080/foo.php?bar=baz#top',
+                'https://example.net:443/foo.php?bar=baz#top',
                 HttpUri::fromString('http://example.net:8080/foo.php?bar=baz#top')
                        ->toHttps()
+                       ->asString()
+        );
+    }
+
+    /**
+     * @since  4.0.2
+     * @test
+     */
+    public function transposingToHttpChangesPort()
+    {
+        $this->assertEquals(
+                'http://example.net:80/foo.php?bar=baz#top',
+                HttpUri::fromString('http://example.net:8080/foo.php?bar=baz#top')
+                       ->toHttp()
                        ->asString()
         );
     }
@@ -292,12 +306,26 @@ class HttpUriTest extends \PHPUnit_Framework_TestCase
      * @since  2.0.0
      * @test
      */
-    public function transposingToHttpLeavesEverythingExceptScheme()
+    public function transposingToHttpLeavesEverythingExceptSchemeAndPort()
     {
         $this->assertEquals(
-                'http://example.net:8080/foo.php?bar=baz#top',
+                'http://example.net:80/foo.php?bar=baz#top',
                 HttpUri::fromString('https://example.net:8080/foo.php?bar=baz#top')
                        ->toHttp()
+                       ->asString()
+        );
+    }
+
+    /**
+     * @since  4.0.2
+     * @test
+     */
+    public function transposingToHttpsWithDifferentPort()
+    {
+        $this->assertEquals(
+                'https://example.net:443/foo.php?bar=baz#top',
+                HttpUri::fromString('https://example.net:8080/foo.php?bar=baz#top')
+                       ->toHttps()
                        ->asString()
         );
     }